provide JPMesh as a library
